+++ src/Valkyrja/Application/Data/Data.php
@@ -13,7 +13,7 @@ declare(strict_types=1);
 
 namespace Valkyrja\Application\Data;
 
-use Valkyrja\Cli\Routing\Data as CliData;
+use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Container\Data as ContainerData;
 use Valkyrja\Event\Data as EventData;
 use Valkyrja\Http\Routing\Data as HttpData;
