+++ src/Valkyrja/Application/Valkyrja.php
@@ -19,7 +19,7 @@ use Valkyrja\Application\Data\Config;
 use Valkyrja\Application\Data\Data;
 use Valkyrja\Application\Exception\RuntimeException;
 use Valkyrja\Application\Support\Component;
-use Valkyrja\Cli\Routing\Data as CliData;
+use Valkyrja\Cli\Routing\Data\Data as CliData;
 use Valkyrja\Container\Data as ContainerData;
 use Valkyrja\Container\Manager\Contract\Container;
 use Valkyrja\Event\Data as EventData;
