+++ src/Valkyrja/Cli/Routing/Data/Data.php
@@ -11,7 +11,7 @@ declare(strict_types=1);
  * file that was distributed with this source code.
  */
 
-namespace Valkyrja\Cli\Routing;
+namespace Valkyrja\Cli\Routing\Data;
 
 /**
  * Class Data.
